Use editor slice for PlateTabBar
