+++ ui/src/components/editor/PlateTabBar/index.js
@@ -1,2 +1,2 @@
-import { PlateTabBar } from './PlateTabBar';
+import { PlateTabBar } from './connect';
 export { PlateTabBar };
